Send notification to project supporters when a project update has been added
related to #809
