+++ src/adapters/notifications/MockNotificationAdapter.ts
@@ -195,6 +195,13 @@ export class MockNotificationAdapter implements NotificationAdapterInterface {
     return Promise.resolve(undefined);
   }
 
+  projectAddedAnUpdate(params: { project: Project }): Promise<void> {
+    logger.info('MockNotificationAdapter projectAddedAnUpdate', {
+      projectSlug: params.project.slug,
+    });
+    return Promise.resolve(undefined);
+  }
+
   donationGetPriceFailed(params: {
     project: Project;
     donationInfo: { txLink: string; reason: string };
