+++ src/adapters/notifications/NotificationAdapterInterface.ts
@@ -62,4 +62,5 @@ export interface NotificationAdapterInterface {
   broadcastNotification(
     params: BroadCastNotificationInputParams,
   ): Promise<void>;
+  projectAddedAnUpdate(params: { project: Project }): Promise<void>;
 }
