+++ src/analytics/analytics.ts
@@ -34,12 +34,11 @@ export enum NOTIFICATIONS_EVENT_NAMES {
   DONATION_RECEIVED = 'Donation received',
   DONATION_GET_PRICE_FAILED = 'Donation get price failed',
   PROJECT_RECEIVED_HEART = 'project liked',
-  PROJECT_UPDATED_DONOR = 'Project updated - donor',
-  PROJECT_UPDATED_OWNER = 'Project updated - owner',
-  PROJECT_UPDATED_USERS_WHO_LIKED = 'Project update - Users Who Liked',
+  PROJECT_UPDATE_ADDED_OWNER = 'Project update added - owner',
   PROJECT_CREATED = 'The project saved as draft',
   UPDATED_PROFILE = 'Updated profile',
   GET_DONATION_PRICE_FAILED = 'Get Donation Price Failed',
   VERIFICATION_FORM_GOT_DRAFT_BY_ADMIN = 'Verification form got draft by admin',
   RAW_HTML_BROADCAST = 'Raw HTML Broadcast',
+  PROJECT_ADD_AN_UPDATE_USERS_WHO_SUPPORT = 'Project update added - Users who supported',
 }
